SSCS Ex v0.1.0 SscsEx

SscsEx: A Simple Sftp Client/Server written in Elixir. See https://hexdocs.pm/elixir/Application.html for more information on OTP Applications

Functions

clean_failed_transfers()

Remove all failed transfers from the list:

Inputs:

None.

Outputs:

"Failed transfers table cleaned up."

display_failed_transfers()

Display the list of failed transfers:

Inputs:

None.

Outputs:

List of transfers

get_conf()

Display the configuration of the SSCS server (system variables SSCS_* values):

Inputs:

None.

Outputs:

List of SSCS_* values

get_file(local_path, remote_path, host, port)

Download of a remote file.

Inputs:

local_path: path of the local file,

remote_path: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

get_file_append(local_path, remote_path, host, port)

Resume the download of a remote file.

Inputs:

local_path: path of the local file,

remote_path: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

ping(host, port)

Ping a remote server:

Inputs:

host: name of remote host,

port: port of remote host.

Outputs:

{:ok, "SSH Host host, port is reachable"} | {:error, "SSH Host host, port is unreachable" }

send_file(local_path, remote_path, host, port)

Upload of a local file.

Inputs:

local_path: path of the local file,

remote_path: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

send_file_append(local_path, remote_path, host, port)

Resume the upload of a local file.

Inputs:

local_path: path of the local file,

remote_path: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

start()

start(type, args)

Called when an application is started.

This function is called when an application is started using Application.start/2 (and functions on top of that, such as Application.ensure_started/2). This function should start the top-level process of the application (which should be the top supervisor of the application's supervision tree if the application follows the OTP design principles around supervision).

start_type defines how the application is started:

:normal - used if the startup is a normal startup or if the application is distributed and is started on the current node because of a failover from another node and the application specification key :start_phases is :undefined.

{:takeover, node} - used if the application is distributed and is started on the current node because of a failover on the node node.

{:failover, node} - used if the application is distributed and is started on the current node because of a failover on node node, and the application specification key :start_phases is not :undefined.

start_args are the arguments passed to the application in the :mod specification key (e.g., mod: {MyApp, [:my_args]}).

This function should either return {:ok, pid} or {:ok, pid, state} if startup is successful. pid should be the PID of the top supervisor. state can be an arbitrary term, and if omitted will default to []; if the application is later stopped, state is passed to the stop/1 callback (see the documentation for the c:stop/1 callback for more information).

use Application provides no default implementation for the start/2 callback.

Callback implementation for Application.start/2.

stop()

"Gracefully" stops the SSCS server (close SSH + save failed transfers on disk):

Inputs:

None.

Outputs:

"Gracefully terminating the SSCS server..."