SSCS Ex v0.1.0 SscsEx.Append

Functions

local_file(connection, local_path, remote_path, checksum_type)

Append the excess part of the local_file to the remote file, accounting their respectives sizes. If the remote file doesn't exist, it is created and the local file is fully downloaded in it. NB: checksum_type could be: :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4 ## Inputs:

connection: SSH connection,

local_path: path of the local file,

remote_path: path of the remote file,

checksum_type: can be :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4

Outputs:

{:ok, "Checksum (type: checksum_type) for both files = checksum"} | {:error, "Couldn't resume uploading local file local_path to remote file remote_path: reason: reason"

local_file_streamed(connection, local_path, remote_path, checksum_type, bytes_size)

Append "streamingly" the excess part of a local file to a remote file, accounting their respectives sizes .. If the remote file doesn't exist, it is created and the local file is fully appended to it (ie uploaded).

Inputs:

connection: SSH connection,

local_path: path of the local file,

remote_path: path of the remote file,

checksum_type: can be :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4,

bytes_size: size of the chunks used in streaming.

Outputs:

{:ok, "Checksum (type: checksum_type) for both files = checksum"} | {:error, "Bad checksum (type: checksum_type}) : local file checksum = local checksum} , remote file checksum = remote checksum"} | {:error, reason}

remote_file(connection, local_path, remote_path, checksum_type)

Append the excess part of the remote_file to the local_file, accounting their respectives sizes.. If the local file doesn't exist, it is created and the remote file is fully downloaded in it.

Inputs:

connection: SSH connection,

local_path: path of the local file,

remote_path: path of the remote file,

checksum_type: can be :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4

Outputs:

{:ok, "Checksum (type: checksum_type) for both files = checksum"},

{:error, "Couldn't resume downloading local file local_path from remote file remote_path: reason: reason"

remote_file_streamed(connection, local_path, remote_path, checksum_type, bytes_size)

Append "streamingly" the excess part of the remote_file to the local_file, accounting their respectives sizes .. If the local file doesn't exist, it is created and the remote file is fully appended to it (ie downloaded).

Inputs:

connection: SSH connection,

local_path: path of the local file,

remote_path: path of the remote file,

checksum_type: can be :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4,

bytes_size: size of the chunks used in streaming.

Outputs:

{:ok, "Checksum (type: checksum_type) for both files = checksum"} | {:error, "Bad checksum (type: checksum_type}) : local file checksum = local checksum} , remote file checksum = remote checksum"} | {:error, reason}