SSCS Ex v0.1.0 SscsEx.Files

Functions to manage local files.

Functions

append_part(path, data)

Append data to the end of local file.

Inputs:

path: path of the file,

data: datas to append.

Outputs:

Returns: {:ok, nil} or {:error, reason}

append_part_from_stream(path, stream_data, bytes_size)

Append stream data to the end of local file.

Inputs:

path: path of the file

stream_data: stream of datas to append,

bytes-size: size of chunks of data to be streamed

Outputs:

{:ok, nil} | {:error, reason}

checksum(file_path, crc_type)

Compute the checksum of a file.

Inputs:

file_path: path of the file,

crc_type: can be :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4

Outputs:

{:ok, <checksum} | {:error, "Wrong file path..."}

create(path)

Create an empty file.

Inputs:

path: path of the file

Outputs:

{:ok, nil} | {:error, reason}

is_dir(strdir)

Check if a path is a folder.

Inputs:

strdir: path of the folder.

Outputs:

:ok | :error

is_file(path)

Check if a file is of the regular type.

Inputs:

path: path of the file.

Outputs:

:ok | :error

read_part(path, start, length)

Read the local file part from position 'start' with the length 'length'.

Inputs:

path: path of the file,

start: start position to read on,

length: size to read

Outputs:

{:ok, data} | {:error, reason}

read_part_to_stream(path, start, bytes_size)

Read the local file part from position 'start' and create a stream from it.

Inputs:

path: path of the file

start: start position to read on,

bytes-size: size of chunks of data to be streamed

Outputs:

{:ok, stream} | {:error, reason}

remove(path)

Remove a file.

Inputs:

path: path of the file

Outputs:

{:ok, "Local file path removed"} | or {:error, "Local file path NOT removed. Reason: reason}"}

size(path)

Give the file size of a file.

Inputs:

strdir: path of the file.

Outputs:

{:ok, size} | {:error, reason}

size2(path)

Give the file size of a file or 0 if the file doesn't exist.

Inputs:

path: path of the file.

Outputs:

{:ok, size} | {:error, "path is not a regular file"} | {:error, "Could not get the size of path. Reason: reason"}}

transfer_rate(path, seconds, unit)

Compute the transfer rate of a file.

Inputs:

path: path of the file,

seconds: transfer duration,

unit: file size unit. Can be: :mb (MB) or :kb (KB)

Outputs:

{:ok, transfer rate} or {:error, message}