SSCS Ex v0.1.0 SscsEx.SFTP

SFTP Functions for the client side of the SscsEx server.

Types

connection_t()

connection_t() :: %{
  channel_pid: pid(),
  connection_ref: :ssh.connection_ref(),
  host: str_or_charlist_t(),
  port: non_neg_integer(),
  opts: keyword()
}

crc_t()

crc_t() ::
  :sha
  | :sha224
  | :sha256
  | :sha384
  | :sha512
  | :sha3_224
  | :sha3_256
  | :sha3_384
  | :sha3_512
  | :blake2b
  | :blake2s
  | :md5
  | :md4

str_or_charlist_t()

str_or_charlist_t() :: charlist() | String.t()

Functions

append_part_file(connection, path, data)

append_part_file(connection_t(), str_or_charlist_t(), iodata()) ::
  {:ok, any()} | {:error, any()}

Append data to the end of remote file.

Inputs:

connection: SFTP connection),

path: path of the file (String),

data: data to append.

Outputs:

{:ok, nil} | {:error, reason}

append_part_file_from_stream(connection, path, stream_data, bytes_size)

append_part_file_from_stream(
  connection_t(),
  str_or_charlist_t(),
  Enumerable.t(),
  non_neg_integer()
) :: {:ok, nil} | {:error, any()}

Append stream data to the end of remote file stream.

Inputs:

connection: SFTP connection,

path: path of the file,

stream_data: datas from stream,

bytes_size: size of stream chunks.

Outputs:

{:ok, nil} | {:error, reason}

close_file_handle(connection, handle)

close_file_handle(connection_t(), any()) :: :ok | {:error, :ssh_sftp.reason()}

Closes an opened remote file.

Inputs:

Connection: SFTP connection,

handle: file handle.

Outputs: :ok | {:error, reason}

connect(opts)

connect(keyword()) ::
  {:ok, connection_t()} | {:error, String.t()} | {:error, any()}

Creates an SFTP channel

Inputs:

opts: SSH options,

Outputs:

{:ok, connection}, or {:error, reason}

create_file(connection, path)

create_file(connection_t(), str_or_charlist_t()) ::
  {:ok, nil} | {:error, any()}

Create an empty remote file given a connection and path. ## Inputs:

Connection: SFTP connection,

path: path of the remote file .

Outputs:

{:ok, nil} | {:error, reason}

disconnect(connection)

disconnect(connection_t()) :: {:error, any()} | :ok

Stops a SFTP channel and closes the SSH connection.

Inputs:

connection: SFTP connection

Output:

:ok | {:error, reason}

each_binstream(connection, handle, byte_length)

each_binstream(connection_t(), any(), non_neg_integer()) ::
  {:halt | [...], any()}

Read a bitstream from a file (handle).

Inputs:

Connection: SFTP connection,

handle: file handle,

byte_length: number of bytes to read

Outputs:

{:halt, handle} | {[data], handle} | {:error, reason}

file_checksum(connection, path, crc_type)

file_checksum(connection_t(), str_or_charlist_t(), crc_t()) ::
  {:error, charlist() | String.t()} | {:ok, binary()}

Return the checksum of a remote file.

Inputs:

connection: SFTP connection),

path: path of the file (String),

crc_type: checksum type = :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4.

Outputs:

{:ok, checksum} | {:error, message}

file_exists?(connection, path)

file_exists?(connection_t(), str_or_charlist_t()) :: boolean()

Gives the type of a remote file given a channel PID and path. ## Inputs:

Connection: SFTP connection,

path: path of the file

Outputs:

(boolean) true | false

file_info(connection, path)

file_info(connection_t(), str_or_charlist_t()) ::
  {:ok, File.Stat.t()} | {:error, any()}

get the file info of a remote file

Inputs:

Connection: SFTP connection,

path: path of the file (String)

Outputs:

{:ok, File.Stat} | {:error, reason}

file_size(connection, path)

file_size(connection_t(), str_or_charlist_t()) ::
  {:ok, non_neg_integer()} | {:error, String.t()} | {:error, any()}

Gives the type of a remote file given a connection and path. ## Inputs:

Connection SFTP connection,

path: path of the remote file.

Outputs:

{:ok, type} | {:error, reason}

file_size2(connection, path)

file_size2(connection_t(), any()) ::
  {:error, String.t()} | {:ok, non_neg_integer()}

Return the size of a remote file or 0 if the file doesn't exist:

Inputs:

connection: SFTP connection),

path: path of the remote file.

Outputs:

{:ok, size} if the remote file exists and is actually a file, | {:ok, 0} if the remote file doesn't exist, | {:error, reason} if the remote file is not a file.

file_type(connection, path)

file_type(connection_t(), str_or_charlist_t()) ::
  {:ok, atom()} | {:error, any()}

Gives the type of a remote file given a channel PID and path. ## Inputs:

Connection: SFTP connection,

path: path of the remote file.

Outputs:

{:ok, type} | {:error, reason}

is_file(connection, path)

is_file(connection_t(), str_or_charlist_t()) :: :error | :ok

Test if a remote file is of the regular type.

Inputs:

Connection: SFTP connection,

path: path of the remote file.

Outputs:

:ok | :error

open_file(connection, path, mode)

open_file(connection_t(), str_or_charlist_t(), [atom()]) ::
  {:ok, any()} | {:error, :ssh_sftp.reason()}

Opens a remote file given a connection and path. ## Inputs:

connection: SFTP connection,

path: path of the remote file,

mode: open modes ([:creat, :write, ...]).

Outputs:

{:ok, handle} | {:error, reason}

read_part_file(connection, path, start, length)

read_part_file(
  connection_t(),
  str_or_charlist_t(),
  non_neg_integer(),
  non_neg_integer()
) :: {:ok, any()} | {:error, any()}

Read 'length' bytes from the remote file part, starting at 'start' position.

Inputs:

connection: SFTP connection),

path: path of the file (String),

start: start position (integer),

length: number of bytes to read (integer)

Outputs:

{:ok, data} | {:error, reason}

read_part_file_to_stream(connection, path, start, bytes_size)

read_part_file_to_stream(
  connection_t(),
  str_or_charlist_t(),
  non_neg_integer(),
  non_neg_integer()
) :: {:ok, any()} | {:error, any()}

Read the remote file part from position 'start' and stream it.

Inputs:

connection: SFTP connection,

path: path of the remote file,

start: start position,

bytes_size: size of stream chunks.

Outputs:

{:ok, stream} | {:error, reason}

remove_file(connection, path)

remove_file(connection_t(), str_or_charlist_t()) ::
  {:error, String.t()} | {:ok, String.t()}

Remove a remote file.

Inputs:

Connection: SFTP connection,

path: path of the remote file.

Outputs:

{:ok, message} or {:error, message}

stream_create!(connection, path, byte_size \\ 32768)

stream_create!(connection_t(), str_or_charlist_t(), non_neg_integer()) :: any()

Open a stream on a file for create and write ## Inputs:

connection: SFTP connection),

path: path of the file (String),

byte_size: stream chunks size (integer)

Outputs:

StreamFile or {:error, reason}

stream_read!(connection, path, byte_size \\ 32768)

stream_read!(connection_t(), str_or_charlist_t(), non_neg_integer()) :: any()

Open a stream on a file for reading ## Inputs:

connection: SFTP connection),

path: path of the file (String),

byte_size: stream chunks size (integer)

Output:

StreamFile | {:error, reason}

stream_write!(connection, path, byte_size \\ 32768)

stream_write!(connection_t(), str_or_charlist_t(), non_neg_integer()) :: any()

Open a stream on a file to write ## Inputs:

connection: SFTP connection),

path: path of the file (String),

byte_size: stream chunks size (integer)

Outputs:

StreamFile | {:error, reason}

write_handle(connection, handle, data)

write_handle(connection_t(), any(), iodata()) ::
  :ok | {:error, :ssh_sftp.reason()}

Writes data to a open file using the channel PID ## Inputs:

Connection: SFTP connection,

handle: handle of the file,

data: data to write

Outputs:

:ok | :error