SSCS Ex v0.1.0 SscsEx.Server

Functions for the server side of the SscsEx server.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

clean_table()

Clean up the content of failed transfers table:

display_table()

Display the content of failed transfers table:

get_file(local_path, remote_path, host, port)

Download of a remote file.

Inputs:

local_file: path of the local file,

remote_file: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

get_file_append(local_path, remote_path, host, port)

Resume the download of a remote file.

Inputs:

local_file: path of the local file,

remote_file: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

handle_continue(atom, map)

Invoked to handle continue instructions.

It is useful for performing work after initialization or for splitting the work in a callback in multiple steps, updating the process state along the way.

Return values are the same as c:handle_cast/2.

This callback is optional. If one is not implemented, the server will fail if a continue instruction is used.

This callback is only supported on Erlang/OTP 21+.

Callback implementation for GenServer.handle_continue/2.

init(options)

Initialize the SSCS server:

Inputs:

_ options: SSH options

Outputs:

{:ok, state} | {:error, reason}

ping(host, port)

Ping a remote server:

Inputs:

host: nale of remote host,

port: port of remote host.

Outputs:

{:ok, "SSH Host host, port is reachable"} | {:error, "SSH Host host, port is unreachable" }

send_file(local_path, remote_path, host, port)

Upload of a local file.

Inputs:

local_file: path of the local file,

remote_file: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

send_file_append(local_path, remote_path, host, port)

Resume the upload of a local file.

Inputs:

local_file: path of the local file,

remote_file: path of the remote file,

host: host name of remote server,

port: port number of listening remote server

Outputs:

{:ok, message} | {:error, reason}

start_link(options)

terminate(reason, state)

(Gracefully) Terminate the SSCS server:

Save the failed transfers database to disk,

Stop the SSH daemon.