SSCS Ex v0.1.0 SscsEx.TransferDb

SscsEx.TransferDb: Manage the ETS/DETS database where the transfers are recorded and the failed ones, kept. When the SSCS server inits, DETS (ie on disk) table is copied to the ETS (ie in memory) table. If there's some failed transfers remaining, they are then retryed. All operation are then done on the ETS table and when (or if) the server terminate, the ETS table is saved back to the DETS table.

Types

t()

t() :: %SscsEx.TransferDb{
  conn_opts: Keyword.t(),
  function: atom(),
  local_path: String.t(),
  remote_path: String.t()
}

Functions

add(table, id, function, local_path, remote_path, conn_opts)

Add a transfer to ETS Table.

Inputs:

table: name of the ETS table,

id: transfer Id,

function: transfer function,

local_path: path of local file,

remote_path: path of remote file,

host: name of remote server,

port: port of remote server

Outputs:

{:ok, nil} | {:error, message}

clean(table)

Remove all transfers from the ETS table.

Inputs:

table: ets table,

Outputs:

{:ok, nil} | {:error, "No transfer table available."}

dets_to_ets(file_name)

Copy DETS file content to ETS Table.

Inputs:

file_name: name of the DETS file.

Outputs:

{:ok, ets table} | {:error, message}

display(table)

Display the ETS table content on standard output.

Inputs:

table: ets table,

Outputs:

{:ok, "Failed transfers table content: transfer 1 , ... , transfer n" } | {:error, "No transfer table available."}

elem_to_list(transfer_db)

ets_to_dets(ets, file_name)

Copy ETS table content to DETS file.

Inputs:

ets: ets table,

file_name: name of the DETS file.

Outputs:

{:ok, nil} | {:error, message}

find(table, id)

find a transfer in the ETS table.

Inputs:

table: ets table,

id: transfer Id.

Outputs:

{:ok, record} | {:error, message}

remove(table, id)

Remove a transfer from the ETS table.

Inputs:

table: ets table,

id: transfer Id.

Outputs:

{:ok, nil} | {:error, message}

tab2list(table)

Get the ETS table content and return it in a list.

Inputs:

table: ets table,

Outputs:

{:ok, [object 1, ... , object n]} | {:error, "No transfer table available."}