SSCS Ex v0.1.0 SscsEx.Utils

Functions

are_options(options, required_keys)

Test if every keys of required_keys are present in the keyword list options. ## Input:

options: keyword list to test,

required_keys: list of keys to test.

Outputs:

{:ok, nil} | {:error, message}

force_tuple_result(any)

Force the :ok/:error functions to return a tuple.

Inputs:

item: :ok, :error or any.

Outputs:

{:ok, nil} | {:error, nil} | any

get_charlist(str)

Force a String to be converted to a charlist (to feed the Erlang functions).

Input:

str: String or charlist.

Outputs:

char list

is_port(num)

Test if num is a valid port number.

Input:

num: an integer.

Outputs:

:ok | :error

log_suppress(fun)

Disable logging while the function 'fun' is executed.

Input:

fun: a function.

Outputs:

The output of the function

stream_checksum(stream, crc_type)

Return the checksum of a stream.

Input:

stream: the stream,

crc_type: type of checksum. can be: :sha, :sha224, :sha256, :sha384, :sha512, :sha3_224, :sha3_256, :sha3_384, :sha3_512, :blake2b, :blake2s, :md5 or :md4

Outputs:

checksum

transfer_rate(file_size, seconds, unit)

Compute a transfer rate. ## Input:

file_size: the size of the file,

seconds: the time elapsed in seconds,

unit: the unit: can be: :mb (MB) or :kb (KB).

Outputs:

{:ok, transfer rate} | {:error, message}

tuple_to_response(arg, default)

Return the second part of a response tuple ({:ok, item} or {:error, item}) when :ok. default otherwise. (Usefull for piping !)

Inputs:

{ok_or_error, item }: {:ok, item} | {:error, item}

default: default value when ok_or_error is not :ok

Outputs:

item or default